refactor: TT-52 package 구조 변경
domain에 따라 package를 생성하고 공통된 코드를 모아놓기 위해 common package를 만듬
