+++ src/main/java/com/twentythree/peech/script/dto/ParagraphResponseDTO.java
@@ -1,4 +1,4 @@
-package com.twentythree.peech.dto;
+package com.twentythree.peech.script.dto;
 
 
 import lombok.Data;
