+++ src/main/java/com/twentythree/peech/script/dto/ScriptRequestDto.java
@@ -1,4 +1,4 @@
-package com.twentythree.peech.dto;
+package com.twentythree.peech.script.dto;
 
 
 public record ScriptRequestDto(String script) {
